+++ src/breadcrumb/index.tsx
@@ -1,2 +1,2 @@
-export { Breadcrumb, BreadcrumbProps, CrumbProps } from './Breadcrumb'
+export { Breadcrumb, BreadcrumbProps, CrumbProps, DividerPosition } from './Breadcrumb'
 export { Breadcrumb as default } from './Breadcrumb'