+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[UPDATE]** Add prop `divider` with `DividerPosition` enum to  `Breadcrumb` 
 - [...]
 
 # v41.9.1 (25/11/2020)